chore: Start with a simple search and web design
The path to the config is hard coded in which is not a good thing

Search:
The search uses recursive line by line naive search.
This is as slow as the computer and the number of files passed to it.

Web design:
The design is not as intuitive as it is supposed to be

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>