chore: Split up files and improve binary check
The files have been split up for easy maintanace.
The previous monolith wasn't large but was dealing with multiple tasks

The binary check and display is now improved.
If the file can be displayed by the browser, the task is left for the browser to do.
If it can't then tell the user so.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>