chore: Improve ui, error display and defaults
The UI is now slightly improved.

Errors e.g 404 are displayed properly instead of a black screen with white text.
This ensures that the user can easily come back from an error (mentally) without being frustrated.

Settings options have been added along with defaults.
The defaults aim to be as good as possible for general usage.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>