chore: Display some media files, improve exprience
media:
Instead of giving the browser the freedom to display media directly,
tell it how.
This is only done in a few files that can be displayed by the browser.
For pdf, just give the browser full control as it knows best!

isBinary:
Add more control codes that can be available in a file.
The previous one was unix centric which means files edited
in Windows will be detected as not text.

ext:
Add a new filter based on file extension e.g `go` or `rs`.
This is to narrow dowm search to only files that you expect.

language:
Use an external tool that is based on github linguist to
detect the current language.
This is not perfect but works better than the previous reliance on `chroma.Lexer`.
Add the number of lines in a file.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>